services/preprocessing/requirements_preprocessing.txt
flask
#flask_restful
#torchvision.transforms 
#PIL 
#io
#cv2
numpy
google.cloud
google-auth

grpcio==1.71
grpcio-tools==1.71
protobuf==5.29
#pydicom
#torchvision
#pillow
opencv-python-headless
google-cloud-storage
torch
torchvision
pillow


